valencia/requirements.txt
Flask
Flask-Cors
tensorflow
Pillow
numpy
Werkzeug
google-genai
python-dotenv
gunicorn
unicorn
websockets
pydantic